+++ core/src/main/scala/spark/ui/Page.scala
@@ -1,0 +1,3 @@
+package spark.ui
+
+private[spark] object Page extends Enumeration { val Storage, Jobs = Value }
